Refactor: Formalise batch notify function module structure
